+++ src/lib.rs
@@ -18,7 +18,7 @@ pub mod mesh;
 pub mod physics;
 
 #[cfg(feature = "rat-mlfmm")]
-pub mod rat_mlfmm;
+pub mod mlfmm;
 
 #[cfg(test)]
 pub(crate) mod testing;